# swagger-specs/swagger-ui-dist/personal-rag-server-openapi.yaml
openapi: 3.0.3
info:
    title: Personal RAG Server API
    description: |
        Personal RAG Server with FastAPI - Secure and Production Ready

        Vector Database: Pinecone (german-philosophic-index-12-worldviews)
        Pinecone Host: https://german-philosophic-index-12-worldviews-ssotzaw.svc.aped-4627-b74a.pinecone.io
    version: 1.0.0
    contact:
        name: API Support
servers:
    - url: /api/v1
      description: API v1
components:
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT
        apiKeyAuth:
            type: apiKey
            in: header
            name: X-API-Key
    schemas:
...
        Assistant:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                description:
                    type: string
                model:
                    type: string
                instructions:
                    type: string
                weltanschauung:
                    type: string
                    enum: [Idealismus, Materialismus, Realismus, Spiritualismus]
                temperature:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 2
                top_p:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                created_at:
                    type: string
                    format: date-time
                updated_at:
                    type: string
                    format: date-time
                status:
                    type: string
                    enum: [Ready, Initializing, Failed]
                metadata:
                    type: object
                    additionalProperties: true
                capabilities:
                    $ref: '#/components/schemas/AssistantCapabilities'
            required:
                - id
                - name
                - weltanschauung
                - model
                - status
        AssistantCapabilities:
            type: object
            properties:
                templates:
                    type: array
                    items:
                        type: string
                        enum: [resolve, reformulate, glossary]
                max_context_length:
                    type: integer
                supported_languages:
                    type: array
                    items:
                        type: string
                        enum: [de, en]
        AssistantListResponse:
            type: object
            properties:
                object:
                    type: string
                    default: 'list'
                    description: Object type
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/Assistant'
                    description: List of assistants
                first_id:
                    type: string
                    nullable: true
                    description: First assistant ID
                last_id:
                    type: string
                    nullable: true
                    description: Last assistant ID
                has_more:
                    type: boolean
                    default: false
                    description: Whether there are more assistants
            required:
                - object
                - data
                - has_more
        AssistantCreate:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                model:
                    type: string
                    default: 'deepseek-reasoner'
                instructions:
                    type: string
                weltanschauung:
                    type: string
                    enum: [Idealismus, Materialismus, Realismus, Spiritualismus]
                temperature:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 2
                    default: 1.0
                top_p:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                    default: 1.0
                metadata:
                    type: object
                    additionalProperties: true
            required:
                - name
                - weltanschauung
                - instructions
        AssistantUpdate:
            type: object
            properties:
                name:
                    type: string
                description:
                    type: string
                instructions:
                    type: string
                temperature:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 2
                top_p:
                    type: number
                    format: float
                    minimum: 0
                    maximum: 1
                metadata:
                    type: object
                    additionalProperties: true
...
paths:
...
    /assistants:
        get:
            summary: List all assistants with optional filtering and capabilities
            operationId: list_assistants
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: weltanschauung
                  in: query
                  required: false
                  schema:
                      type: string
                      enum:
                          [Idealismus, Materialismus, Realismus, Spiritualismus]
                  description: Filter assistants by weltanschauung
                  example: Idealismus
                - name: status
                  in: query
                  required: false
                  schema:
                      type: string
                      enum: [Ready, Initializing, Failed]
                  description: Filter assistants by status
                  example: Ready
                - name: include_capabilities
                  in: query
                  required: false
                  schema:
                      type: boolean
                      default: false
                  description: Include template capabilities in response
                  example: true
            responses:
                '200':
                    description: List of assistants with capabilities
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AssistantListResponse'
                            example:
                                assistants:
                                    - id: 'idealismus-aurelian-i-schelling'
                                      name: 'Aurelian I. Schelling'
                                      weltanschauung: 'Idealismus'
                                      model: 'deepseek-reasoner'
                                      status: 'Ready'
                                      description: 'Philosophical assistant for Idealism worldview'
                                      created_at: '2024-01-15T10:30:00Z'
                                      capabilities:
                                          templates:
                                              [
                                                  'resolve',
                                                  'reformulate',
                                                  'glossary',
                                              ]
                                          max_context_length: 8192
                                          supported_languages: ['de', 'en']
                                total_count: 4
                                available_worldviews:
                                    [
                                        'Idealismus',
                                        'Materialismus',
                                        'Realismus',
                                        'Spiritualismus',
                                    ]
                '401':
                    description: Unauthorized

        post:
            summary: Create a new assistant
            operationId: create_assistant
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AssistantCreate'
            responses:
                '201':
                    description: Assistant created successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Assistant'
                '400':
                    description: Invalid request
                '401':
                    description: Unauthorized
                '500':
                    description: Failed to create assistant
    /assistants/{assistant_id}:
        get:
            summary: Get assistant details
            operationId: get_assistant
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: assistant_id
                  in: path
                  required: true
                  schema:
                      type: string
                  example: 'idealismus-aurelian-i-schelling'
            responses:
                '200':
                    description: Assistant details
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Assistant'
                '401':
                    description: Unauthorized
                '404':
                    description: Assistant not found

        put:
            summary: Update an assistant
            operationId: update_assistant
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: assistant_id
                  in: path
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AssistantUpdate'
            responses:
                '200':
                    description: Assistant updated successfully
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Assistant'
                '400':
                    description: Invalid request
                '401':
                    description: Unauthorized
                '404':
                    description: Assistant not found
                '500':
                    description: Failed to update assistant

        delete:
            summary: Delete an assistant
            operationId: delete_assistant
            tags:
                - assistants
            security:
                - bearerAuth: []
                - apiKeyAuth: []
            parameters:
                - name: assistant_id
                  in: path
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Assistant deleted successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string
                '401':
                    description: Unauthorized
                '404':
                    description: Assistant not found
                '500':
                    description: Failed to delete assistant
...